+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc16072/OpModes/QQOpMode.java
@@ -16,4 +16,9 @@ public abstract class QQOpMode extends OpMode {
     public void init() {
         robot.init(hardwareMap);
     }
+
+    @Override
+    public void loop(){
+        robot.lift.update();
+    }
 }
