+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc16072/OpModes/StackTestAuto.java
@@ -2,15 +2,13 @@ package org.firstinspires.ftc.teamcode.ftc16072.OpModes;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
-import org.firstinspires.ftc.robotcore.internal.android.dx.ssa.DomFront;
 import org.firstinspires.ftc.teamcode.ftc16072.actions.ChangeLiftAction;
-import org.firstinspires.ftc.teamcode.ftc16072.actions.HorizontalSlides;
+import org.firstinspires.ftc.teamcode.ftc16072.mechanisms.Lift;
 
 @Autonomous
 public class StackTestAuto extends VisionAutoBase {
     @Override
     public void start(){
-        return new ChangeLiftAction("move lift up", 400 )
-                .setNext(HorizontalSlides("Move slides out", ))
+        currentAction = new ChangeLiftAction("move lift up", Lift.Level.CONE_FIVE_STACK);
     }
 }
